Clean up and updates to address several pull request comments.
Jira ticket: CAMS-270
